Package as 2.7.0.
