+++ src/Stott.Optimizely.RobotsHandler/Stott.Optimizely.RobotsHandler.csproj
@@ -2,21 +2,21 @@
 
     <PropertyGroup>
         <TargetFramework>net6.0</TargetFramework>
-        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+        <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
         <AddRazorSupportForMvc>true</AddRazorSupportForMvc>
         <Authors>Mark Stott</Authors>
         <PackageProjectUrl>https://github.com/GeekInTheNorth/Stott.Optimizely.RobotsHandler</PackageProjectUrl>
         <RepositoryUrl>https://github.com/GeekInTheNorth/Stott.Optimizely.RobotsHandler</RepositoryUrl>
-        <Copyright>Mark Stott 2023</Copyright>
+        <Copyright>Mark Stott 2024</Copyright>
         <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
         <Description>Provides an Admin interface in Optimizely CMS 12 for managing robots.txt files.  A controller responds to the /robots.txt path and returns the appropriate content specific to the site.</Description>
         <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
         <NeutralLanguage>en-GB</NeutralLanguage>
         <PackageTags>Optimizely CMS Robots Robots_txt</PackageTags>
-        <AssemblyVersion>2.6.2.0</AssemblyVersion>
-        <FileVersion>2.6.2.0</FileVersion>
-        <Version>2.6.2</Version>
-        <PackageReleaseNotes>Prevent redirect to include trailing slash from preventing saving of the robots.txt.</PackageReleaseNotes>
+        <AssemblyVersion>2.7.0.0</AssemblyVersion>
+        <FileVersion>2.7.0.0</FileVersion>
+        <Version>2.7.0</Version>
+        <PackageReleaseNotes>UI Rebuild on vite and React.</PackageReleaseNotes>
         <PackageReadmeFile>README.md</PackageReadmeFile>
         <RepositoryType>git</RepositoryType>
     </PropertyGroup>
